web/requirements.txt
click==6.7
Flask==0.12
gunicorn==19.7.0
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==1.0
Werkzeug==0.12
Flask-DebugToolbar==0.10.1
flask-assets
flask_sqlalchemy
markdown
Pygments
bcrypt
flask-admin
flask-login
Flask-WTF
WTForms
cssmin
psycopg2
python-u2flib-server